--- a/src/hotspot/cpu/s390/c1_Runtime1_s390.cpp
+++ b/src/hotspot/cpu/s390/c1_Runtime1_s390.cpp
@@ -589,6 +589,79 @@ OopMapSet* Runtime1::generate_code_for(C1StubId id, StubAssembler* sasm) {
       __ z_br(Z_R14);
     }
     break;
+    case C1StubId::is_instance_of_id:
+    {
+      // Mirror: Z_ARG1(R2)
+      // Object: Z_ARG2
+      // Temps: Z_ARG3, Z_ARG4, Z_ARG5, Z_tmp_1, Z_tmp_2
+      // Result: Z_RET(R2)
+
+      // Get the Klass* into Z_ARG3
+      Register klass = Z_ARG3 , obj = Z_ARG2, result = Z_RET;
+      Register temp0 = Z_ARG4, temp1 = Z_ARG5, temp2 = Z_tmp_1, temp3 = Z_tmp_2;
+
+      __ z_lg(klass, Address(Z_ARG1, java_lang_Class::klass_offset()));
+
+      Label done, is_secondary, same;
+
+      __ clear_reg(result /* Z_R2 */, true /* whole_reg */, false /* set_cc */);  // sets r_result=0 (failure)
+
+      __ z_ltgr(klass, klass); // Klass is null
+      __ branch_optimized(Assembler::bcondEqual, done);
+
+      __ z_ltgr(obj, obj); // obj is null
+      __ branch_optimized(Assembler::bcondEqual, done);
+
+      __ z_llgf(temp0, Address(klass, in_bytes(Klass::super_check_offset_offset())));
+      __ compare32_and_branch(temp0, in_bytes(Klass::secondary_super_cache_offset()), Assembler::bcondEqual, is_secondary); // Klass is a secondary superclass
+
+      // Klass is a concrete class
+      __ load_klass(temp1, obj);
+      __ z_cg(klass, Address(temp1, temp0));
+
+      // result is already holding 0, denoting NotEqual case
+      __ z_lochi(result, 1, Assembler::bcondEqual);
+      __ z_br(Z_R14);
+
+      __ bind(is_secondary);
+
+      __ load_klass(obj, obj);
+
+      // This is necessary because I am never in my own secondary_super list.
+      __ z_cgr(obj, klass);
+      __ branch_optimized(Assembler::bcondEqual, same);
+
+      const int frame_size = 2*BytesPerWord + frame::z_abi_160_size;
+      __ save_return_pc();
+      __ push_frame(frame_size);
+      int i = 0;
+      __ z_stg(Z_tmp_1, (i++)*BytesPerWord + frame::z_abi_160_size, Z_SP);
+      __ z_stg(Z_tmp_2, (i++)*BytesPerWord + frame::z_abi_160_size, Z_SP);
+      assert(i*BytesPerWord + frame::z_abi_160_size == frame_size, "check");
+
+      __ lookup_secondary_supers_table_var(obj, klass,
+                                          /*temps*/ temp0, temp1, temp2, temp3,
+                                          result);
+
+      // lookup_secondary_supers_table_var return 0 on success and 1 on failure.
+      // but this method returns 0 on failure and 1 on success.
+      // so we have to inverse the result we got from lookup_secondary_supers_table_var.
+      __ z_xilf(result, 1);  // inverse the result
+
+      i = 0;
+      __ z_lg(Z_tmp_1, (i++)*BytesPerWord + frame::z_abi_160_size, Z_SP);
+      __ z_lg(Z_tmp_2, (i++)*BytesPerWord + frame::z_abi_160_size, Z_SP);
+      assert(i*BytesPerWord + frame::z_abi_160_size == frame_size, "check");
+      __ pop_frame();
+      __ restore_return_pc();
+      __ z_br(Z_R14); // we have correct result in r2, so let's branch from right here.
+
+      __ bind(same);
+      __ z_lochi(result, 1, Assembler::bcondEqual);
+
+      __ bind(done);
+      __ z_br(Z_R14);
+    }
     case C1StubId::monitorenter_nofpu_id:
     case C1StubId::monitorenter_id:
       { // Z_R1_scratch : object