s390x implementation for Class.isInstance intrinsic. 

Tier1 test on release & fastdebug vm are clean with flag: `-XX:-UseSecondarySupersCache -XX:+UseSecondarySupersTable -XX:+VerifySecondarySupers -XX:+StressSecondarySupers`. 

Benchmark results will be updated soon.