@offamitkumar 2025-02-10
without patch: 

```
Benchmark                             Mode  Cnt   Score   Error  Units
SecondarySupersLookup.testNegative00  avgt   15   1.430 ± 0.106  ns/op
SecondarySupersLookup.testNegative01  avgt   15   1.390 ± 0.085  ns/op
SecondarySupersLookup.testNegative02  avgt   15   1.392 ± 0.086  ns/op
SecondarySupersLookup.testNegative03  avgt   15   1.354 ± 0.011  ns/op
SecondarySupersLookup.testNegative04  avgt   15   1.353 ± 0.009  ns/op
SecondarySupersLookup.testNegative05  avgt   15   1.430 ± 0.104  ns/op
SecondarySupersLookup.testNegative06  avgt   15   1.352 ± 0.008  ns/op
SecondarySupersLookup.testNegative07  avgt   15   1.353 ± 0.008  ns/op
SecondarySupersLookup.testNegative08  avgt   15   1.352 ± 0.009  ns/op
SecondarySupersLookup.testNegative09  avgt   15   1.355 ± 0.012  ns/op
SecondarySupersLookup.testNegative10  avgt   15   1.391 ± 0.086  ns/op
SecondarySupersLookup.testNegative16  avgt   15   1.352 ± 0.008  ns/op
SecondarySupersLookup.testNegative20  avgt   15   1.391 ± 0.086  ns/op
SecondarySupersLookup.testNegative30  avgt   15   1.391 ± 0.086  ns/op
SecondarySupersLookup.testNegative32  avgt   15   1.430 ± 0.104  ns/op
SecondarySupersLookup.testNegative40  avgt   15   1.353 ± 0.010  ns/op
SecondarySupersLookup.testNegative50  avgt   15   1.390 ± 0.087  ns/op
SecondarySupersLookup.testNegative55  avgt   15  25.403 ± 1.269  ns/op
SecondarySupersLookup.testNegative56  avgt   15  26.406 ± 2.156  ns/op
SecondarySupersLookup.testNegative57  avgt   15  26.495 ± 1.960  ns/op
SecondarySupersLookup.testNegative58  avgt   15  28.065 ± 3.160  ns/op
SecondarySupersLookup.testNegative59  avgt   15  28.189 ± 3.006  ns/op
SecondarySupersLookup.testNegative60  avgt   15  29.089 ± 3.349  ns/op
SecondarySupersLookup.testNegative61  avgt   15  27.718 ± 1.070  ns/op
SecondarySupersLookup.testNegative62  avgt   15  28.047 ± 1.146  ns/op
SecondarySupersLookup.testNegative63  avgt   15  28.695 ± 1.611  ns/op
SecondarySupersLookup.testNegative64  avgt   15  29.326 ± 2.023  ns/op
SecondarySupersLookup.testPositive01  avgt   15   1.719 ± 0.049  ns/op
SecondarySupersLookup.testPositive02  avgt   15   1.744 ± 0.080  ns/op
SecondarySupersLookup.testPositive03  avgt   15   1.743 ± 0.076  ns/op
SecondarySupersLookup.testPositive04  avgt   15   1.764 ± 0.087  ns/op
SecondarySupersLookup.testPositive05  avgt   15   1.763 ± 0.085  ns/op
SecondarySupersLookup.testPositive06  avgt   15   1.741 ± 0.075  ns/op
SecondarySupersLookup.testPositive07  avgt   15   1.719 ± 0.050  ns/op
SecondarySupersLookup.testPositive08  avgt   15   1.719 ± 0.049  ns/op
SecondarySupersLookup.testPositive09  avgt   15   1.720 ± 0.050  ns/op
SecondarySupersLookup.testPositive10  avgt   15   1.765 ± 0.087  ns/op
SecondarySupersLookup.testPositive16  avgt   15   1.744 ± 0.076  ns/op
SecondarySupersLookup.testPositive20  avgt   15   1.721 ± 0.053  ns/op
SecondarySupersLookup.testPositive30  avgt   15   1.719 ± 0.051  ns/op
SecondarySupersLookup.testPositive32  avgt   15   1.721 ± 0.052  ns/op
SecondarySupersLookup.testPositive40  avgt   15  12.798 ± 0.150  ns/op
SecondarySupersLookup.testPositive50  avgt   15   1.744 ± 0.076  ns/op
SecondarySupersLookup.testPositive60  avgt   15  24.580 ± 0.567  ns/op
SecondarySupersLookup.testPositive63  avgt   15  23.523 ± 1.635  ns/op
SecondarySupersLookup.testPositive64  avgt   15  33.512 ± 3.343  ns/op
```

with patch:

```
SecondarySupersLookup.testNegative00  avgt   15   1.354 ± 0.009  ns/op
SecondarySupersLookup.testNegative01  avgt   15   1.399 ± 0.086  ns/op
SecondarySupersLookup.testNegative02  avgt   15   1.365 ± 0.054  ns/op
SecondarySupersLookup.testNegative03  avgt   15   1.351 ± 0.008  ns/op
SecondarySupersLookup.testNegative04  avgt   15   1.353 ± 0.010  ns/op
SecondarySupersLookup.testNegative05  avgt   15   1.353 ± 0.009  ns/op
SecondarySupersLookup.testNegative06  avgt   15   1.470 ± 0.106  ns/op
SecondarySupersLookup.testNegative07  avgt   15   1.365 ± 0.055  ns/op
SecondarySupersLookup.testNegative08  avgt   15   1.352 ± 0.008  ns/op
SecondarySupersLookup.testNegative09  avgt   15   1.431 ± 0.106  ns/op
SecondarySupersLookup.testNegative10  avgt   15   1.355 ± 0.012  ns/op
SecondarySupersLookup.testNegative16  avgt   15   1.430 ± 0.107  ns/op
SecondarySupersLookup.testNegative20  avgt   15   1.352 ± 0.008  ns/op
SecondarySupersLookup.testNegative30  avgt   15   1.354 ± 0.009  ns/op
SecondarySupersLookup.testNegative32  avgt   15   1.391 ± 0.084  ns/op
SecondarySupersLookup.testNegative40  avgt   15   1.392 ± 0.086  ns/op
SecondarySupersLookup.testNegative50  avgt   15   1.354 ± 0.011  ns/op
SecondarySupersLookup.testNegative55  avgt   15  25.587 ± 1.993  ns/op
SecondarySupersLookup.testNegative56  avgt   15  26.048 ± 1.970  ns/op
SecondarySupersLookup.testNegative57  avgt   15  27.874 ± 3.353  ns/op
SecondarySupersLookup.testNegative58  avgt   15  26.392 ± 1.136  ns/op
SecondarySupersLookup.testNegative59  avgt   15  26.593 ± 0.393  ns/op
SecondarySupersLookup.testNegative60  avgt   15  27.567 ± 1.394  ns/op
SecondarySupersLookup.testNegative61  avgt   15  28.813 ± 2.429  ns/op
SecondarySupersLookup.testNegative62  avgt   15  28.523 ± 1.723  ns/op
SecondarySupersLookup.testNegative63  avgt   15  29.832 ± 2.802  ns/op
SecondarySupersLookup.testNegative64  avgt   15  30.048 ± 2.749  ns/op
SecondarySupersLookup.testPositive01  avgt   15   1.788 ± 0.092  ns/op
SecondarySupersLookup.testPositive02  avgt   15   1.742 ± 0.076  ns/op
SecondarySupersLookup.testPositive03  avgt   15   1.741 ± 0.074  ns/op
SecondarySupersLookup.testPositive04  avgt   15   1.764 ± 0.084  ns/op
SecondarySupersLookup.testPositive05  avgt   15   1.741 ± 0.076  ns/op
SecondarySupersLookup.testPositive06  avgt   15   1.720 ± 0.051  ns/op
SecondarySupersLookup.testPositive07  avgt   15   1.722 ± 0.054  ns/op
SecondarySupersLookup.testPositive08  avgt   15   1.721 ± 0.053  ns/op
SecondarySupersLookup.testPositive09  avgt   15   1.764 ± 0.086  ns/op
SecondarySupersLookup.testPositive10  avgt   15   1.719 ± 0.050  ns/op
SecondarySupersLookup.testPositive16  avgt   15   1.741 ± 0.074  ns/op
SecondarySupersLookup.testPositive20  avgt   15   1.721 ± 0.053  ns/op
SecondarySupersLookup.testPositive30  avgt   15   1.759 ± 0.111  ns/op
SecondarySupersLookup.testPositive32  avgt   15   1.741 ± 0.074  ns/op
SecondarySupersLookup.testPositive40  avgt   15  12.722 ± 0.085  ns/op
SecondarySupersLookup.testPositive50  avgt   15   1.735 ± 0.098  ns/op
SecondarySupersLookup.testPositive60  avgt   15  25.368 ± 1.966  ns/op
SecondarySupersLookup.testPositive63  avgt   15  23.529 ± 1.987  ns/op
SecondarySupersLookup.testPositive64  avgt   15  30.335 ± 3.803  ns/op
```