@offamitkumar 2025-02-10
command : `make test TEST="micro:vm.lang.SecondarySupersLookup" MICRO=" JAVA_OPTIONS=-XX:+UnlockDiagnosticVMOptions -XX:-UseSecondarySupersCache -XX:TieredStopAtLevel=1"` 

without patch: 

```
Benchmark                             Mode  Cnt   Score   Error  Units
SecondarySupersLookup.testNegative00  avgt   15   6.554 ± 0.023  ns/op
SecondarySupersLookup.testNegative01  avgt   15   6.690 ± 0.428  ns/op
SecondarySupersLookup.testNegative02  avgt   15   6.561 ± 0.019  ns/op
SecondarySupersLookup.testNegative03  avgt   15   6.545 ± 0.004  ns/op
SecondarySupersLookup.testNegative04  avgt   15   6.549 ± 0.011  ns/op
SecondarySupersLookup.testNegative05  avgt   15   6.554 ± 0.027  ns/op
SecondarySupersLookup.testNegative06  avgt   15   6.551 ± 0.019  ns/op
SecondarySupersLookup.testNegative07  avgt   15   6.548 ± 0.009  ns/op
SecondarySupersLookup.testNegative08  avgt   15   6.549 ± 0.015  ns/op
SecondarySupersLookup.testNegative09  avgt   15   6.550 ± 0.014  ns/op
SecondarySupersLookup.testNegative10  avgt   15   6.546 ± 0.004  ns/op
SecondarySupersLookup.testNegative16  avgt   15   6.552 ± 0.014  ns/op
SecondarySupersLookup.testNegative20  avgt   15   6.551 ± 0.017  ns/op
SecondarySupersLookup.testNegative30  avgt   15   6.546 ± 0.006  ns/op
SecondarySupersLookup.testNegative32  avgt   15   6.545 ± 0.002  ns/op
SecondarySupersLookup.testNegative40  avgt   15   6.553 ± 0.016  ns/op
SecondarySupersLookup.testNegative50  avgt   15   6.549 ± 0.012  ns/op
SecondarySupersLookup.testNegative55  avgt   15  16.530 ± 0.043  ns/op
SecondarySupersLookup.testNegative56  avgt   15  16.520 ± 0.030  ns/op
SecondarySupersLookup.testNegative57  avgt   15  16.522 ± 0.036  ns/op
SecondarySupersLookup.testNegative58  avgt   15  16.517 ± 0.028  ns/op
SecondarySupersLookup.testNegative59  avgt   15  19.802 ± 0.298  ns/op
SecondarySupersLookup.testNegative60  avgt   15  21.237 ± 0.044  ns/op
SecondarySupersLookup.testNegative61  avgt   15  21.241 ± 0.050  ns/op
SecondarySupersLookup.testNegative62  avgt   15  21.243 ± 0.042  ns/op
SecondarySupersLookup.testNegative63  avgt   15  25.421 ± 0.033  ns/op
SecondarySupersLookup.testNegative64  avgt   15  25.064 ± 0.089  ns/op
SecondarySupersLookup.testPositive01  avgt   15   9.818 ± 0.026  ns/op
SecondarySupersLookup.testPositive02  avgt   15   9.819 ± 0.017  ns/op
SecondarySupersLookup.testPositive03  avgt   15   9.826 ± 0.025  ns/op
SecondarySupersLookup.testPositive04  avgt   15   9.817 ± 0.019  ns/op
SecondarySupersLookup.testPositive05  avgt   15   9.815 ± 0.022  ns/op
SecondarySupersLookup.testPositive06  avgt   15   9.821 ± 0.018  ns/op
SecondarySupersLookup.testPositive07  avgt   15   9.824 ± 0.035  ns/op
SecondarySupersLookup.testPositive08  avgt   15   9.837 ± 0.041  ns/op
SecondarySupersLookup.testPositive09  avgt   15   9.820 ± 0.030  ns/op
SecondarySupersLookup.testPositive10  avgt   15   9.817 ± 0.008  ns/op
SecondarySupersLookup.testPositive16  avgt   15   9.819 ± 0.016  ns/op
SecondarySupersLookup.testPositive20  avgt   15   9.818 ± 0.012  ns/op
SecondarySupersLookup.testPositive30  avgt   15   9.820 ± 0.013  ns/op
SecondarySupersLookup.testPositive32  avgt   15   9.820 ± 0.024  ns/op
SecondarySupersLookup.testPositive40  avgt   15  12.722 ± 0.029  ns/op
SecondarySupersLookup.testPositive50  avgt   15   9.820 ± 0.020  ns/op
SecondarySupersLookup.testPositive60  avgt   15  12.717 ± 0.015  ns/op
SecondarySupersLookup.testPositive63  avgt   15  22.316 ± 0.024  ns/op
SecondarySupersLookup.testPositive64  avgt   15  24.904 ± 0.057  ns/op
Finished running test 'micro:vm.lang.SecondarySupersLookup'
```

with the patch: 

```
Benchmark                             Mode  Cnt   Score   Error  Units
SecondarySupersLookup.testNegative00  avgt   15   4.780 ± 0.177  ns/op
SecondarySupersLookup.testNegative01  avgt   15   4.719 ± 0.010  ns/op
SecondarySupersLookup.testNegative02  avgt   15   4.766 ± 0.179  ns/op
SecondarySupersLookup.testNegative03  avgt   15   4.723 ± 0.013  ns/op
SecondarySupersLookup.testNegative04  avgt   15   4.761 ± 0.169  ns/op
SecondarySupersLookup.testNegative05  avgt   15   4.760 ± 0.171  ns/op
SecondarySupersLookup.testNegative06  avgt   15   4.719 ± 0.008  ns/op
SecondarySupersLookup.testNegative07  avgt   15   4.719 ± 0.009  ns/op
SecondarySupersLookup.testNegative08  avgt   15   4.718 ± 0.007  ns/op
SecondarySupersLookup.testNegative09  avgt   15   4.761 ± 0.168  ns/op
SecondarySupersLookup.testNegative10  avgt   15   4.762 ± 0.091  ns/op
SecondarySupersLookup.testNegative16  avgt   15   4.719 ± 0.009  ns/op
SecondarySupersLookup.testNegative20  avgt   15   4.721 ± 0.013  ns/op
SecondarySupersLookup.testNegative30  avgt   15   4.762 ± 0.184  ns/op
SecondarySupersLookup.testNegative32  avgt   15   4.884 ± 0.301  ns/op
SecondarySupersLookup.testNegative40  avgt   15   4.721 ± 0.013  ns/op
SecondarySupersLookup.testNegative50  avgt   15   4.719 ± 0.009  ns/op
SecondarySupersLookup.testNegative55  avgt   15  29.569 ± 3.057  ns/op
SecondarySupersLookup.testNegative56  avgt   15  29.835 ± 2.460  ns/op
SecondarySupersLookup.testNegative57  avgt   15  33.406 ± 3.634  ns/op
SecondarySupersLookup.testNegative58  avgt   15  31.665 ± 3.438  ns/op
SecondarySupersLookup.testNegative59  avgt   15  35.713 ± 3.282  ns/op
SecondarySupersLookup.testNegative60  avgt   15  31.220 ± 2.361  ns/op
SecondarySupersLookup.testNegative61  avgt   15  34.202 ± 3.560  ns/op
SecondarySupersLookup.testNegative62  avgt   15  32.143 ± 2.823  ns/op
SecondarySupersLookup.testNegative63  avgt   15  32.445 ± 2.387  ns/op
SecondarySupersLookup.testNegative64  avgt   15  35.546 ± 3.793  ns/op
SecondarySupersLookup.testPositive01  avgt   15   5.211 ± 0.011  ns/op
SecondarySupersLookup.testPositive02  avgt   15   5.225 ± 0.073  ns/op
SecondarySupersLookup.testPositive03  avgt   15   5.211 ± 0.009  ns/op
SecondarySupersLookup.testPositive04  avgt   15   5.211 ± 0.009  ns/op
SecondarySupersLookup.testPositive05  avgt   15   5.228 ± 0.072  ns/op
SecondarySupersLookup.testPositive06  avgt   15   5.852 ± 1.266  ns/op
SecondarySupersLookup.testPositive07  avgt   15   5.213 ± 0.012  ns/op
SecondarySupersLookup.testPositive08  avgt   15   5.234 ± 0.101  ns/op
SecondarySupersLookup.testPositive09  avgt   15   5.227 ± 0.067  ns/op
SecondarySupersLookup.testPositive10  avgt   15   5.214 ± 0.015  ns/op
SecondarySupersLookup.testPositive16  avgt   15   5.213 ± 0.018  ns/op
SecondarySupersLookup.testPositive20  avgt   15   5.209 ± 0.009  ns/op
SecondarySupersLookup.testPositive30  avgt   15   5.208 ± 0.004  ns/op
SecondarySupersLookup.testPositive32  avgt   15   5.266 ± 0.121  ns/op
SecondarySupersLookup.testPositive40  avgt   15  16.094 ± 0.621  ns/op
SecondarySupersLookup.testPositive50  avgt   15   5.215 ± 0.016  ns/op
SecondarySupersLookup.testPositive60  avgt   15  29.342 ± 3.571  ns/op
SecondarySupersLookup.testPositive63  avgt   15  27.752 ± 3.543  ns/op
SecondarySupersLookup.testPositive64  avgt   15  36.571 ± 4.141  ns/op
Finished running test 'micro:vm.lang.SecondarySupersLookup'
```